EBPF Backend: Support bit-string extract & lookahead
Add support for direct extraction and lookahead of bit-strings; this was
a language feature missing from the backend.  The main approach is to
treat the bit-string as a struct with a single field.  Or rather
abstract the handling of structs such that they generate a list of
fields (types + names) and the bit-string is just a simple special case
(single field with the type directly specified and no name).

This requires some adjustments of prototypes here and there, but nothign
significant.  Next step is to clean up the generated code some more.